documentation: UML: correct hostfs mounting
Update the hostfs mounting documentation to reflect commandlines that
currently work and document the hostfs kernel parameter in the same
section.

The hostfs driver ignores almost all mount options and gets the source
directory by concatenating any argument to the hostfs kernel commandline
parameter and the source argument to mount(2). The existing docs
indicate that the source argument is ignored, and the host-side path
should be passed in the mount options.

Signed-off-by: David Finkel <davidf@vimeo.com>